add component to main page (when no tasks)
