+++ frontend/src/features/main/ui/Main.tsx
@@ -1,23 +1,23 @@
-import { WarningCard } from '../warning-card'
-
+import { Calendar } from './calendar/Calendar'
 import s from './main.module.scss'
-import { AssignedCourses } from './tasks-elements/AssignedCourses'
 import { CurrentTasks } from './tasks-elements/CurrentTasks'
 import { ExpiredTasks } from './tasks-elements/ExpiredTasks'
+import { WarningCard } from './warning-card'
 
 export const Main = () => {
+  const hasWarning = true
   return (
     <div className={s.contentBlock}>
       <div className={s.tasks}>
         <CurrentTasks />
         <ExpiredTasks />
       </div>
-      <div className={s.warning}>
-        <WarningCard />
-      </div>
-      <div className={s.assignedCourses}>
-        <AssignedCourses />
-      </div>
+      {hasWarning && (
+        <div className={s.warning}>
+          <WarningCard />
+        </div>
+      )}
+      <Calendar />
     </div>
   )
 }
