+++ frontend/src/features/main/ui/calendar/Calendar.tsx
@@ -1,0 +1,11 @@
+import { Typography } from '@shared/components'
+
+import s from './calendar.module.scss'
+
+export const Calendar = () => {
+  return (
+    <div className={s.calendar}>
+      <Typography variant="body_1">Здесь будет календарь</Typography>
+    </div>
+  )
+}
