+++ frontend/src/features/main/ui/calendar/calendar.module.scss
@@ -1,0 +1,11 @@
+.calendar {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    height: 100px;
+    width: 100%;
+    background-color: var(--color-bg-gray100);
+    color: var(--color-text-primary);
+    border-radius: 8px;
+    margin-bottom: 40px;
+}
