--- frontend/src/features/main/ui/main/index.ts
@@ -1,1 +1,0 @@
-export * from './Main'
