--- frontend/src/features/main/ui/main/tasks-elements/AssignedCourses.tsx
@@ -1,17 +1,0 @@
-import { Typography } from '@shared/components'
-
-import { Task } from '../../task'
-import s from '../main.module.scss'
-
-export const AssignedCourses = () => {
-  return (
-    <>
-      <Typography variant="header_4" className={s.title}>
-        Назначенные курсы
-      </Typography>
-      <Task daysLeft={7} className={s.appointedTask}>
-        Инструкции по эвакуации и действиям в случае ЧС
-      </Task>
-    </>
-  )
-}
