+++ frontend/src/pages/auth/authpage.module.scss
@@ -8,6 +8,10 @@
   max-width: 1350px;
   margin: 0 auto;
   position: relative;
+
+  @media screen and (max-width: 767px){
+    justify-content: center;
+  }
   
 }
